remove inheritance and simplify extensions
