+++ data/crac/crac-api/src/main/java/com/farao_community/farao/data/crac_api/NetworkAction.java
@@ -19,7 +19,7 @@ import com.powsybl.iidm.network.Network;
  * @author Joris Mancini {@literal <joris.mancini at rte-france.com>}
  */
 @JsonTypeInfo(use = JsonTypeInfo.Id.NAME, property = "type")
-public interface NetworkAction extends RemedialAction<NetworkAction> {
+public interface NetworkAction extends RemedialAction {
 
     /**
      * Trigger the actions on the NetworkElements, in a given network.
