+++ data/crac/crac-impl/src/main/java/com/farao_community/farao/data/crac_impl/AbstractRemedialAction.java
@@ -39,7 +39,7 @@ import java.util.List;
         @JsonSubTypes.Type(value = Countertrading.class, name = "countertrading"),
         @JsonSubTypes.Type(value = AlignedRangeAction.class, name = "aligned-range-action")
     })
-public abstract class AbstractRemedialAction<I extends RemedialAction<I>> extends AbstractIdentifiable<I> implements RemedialAction<I> {
+public abstract class AbstractRemedialAction extends AbstractIdentifiable implements RemedialAction {
     protected String operator;
     protected List<UsageRule> usageRules;
