+++ data/crac/crac-impl/src/main/java/com/farao_community/farao/data/crac_impl/remedial_action/range_action/AbstractElementaryRangeAction.java
@@ -38,7 +38,7 @@ import java.util.stream.Collectors;
         @JsonSubTypes.Type(value = InjectionRange.class, name = "injection-range"),
         @JsonSubTypes.Type(value = Redispatching.class, name = "redispatching")
     })
-public abstract class AbstractElementaryRangeAction extends AbstractRemedialAction<RangeAction> implements RangeAction {
+public abstract class AbstractElementaryRangeAction extends AbstractRemedialAction implements RangeAction {
     protected List<Range> ranges;
     protected NetworkElement networkElement;
