+++ data/crac/crac-impl/src/main/java/com/farao_community/farao/data/crac_impl/remedial_action/range_action/AlignedRangeAction.java
@@ -29,7 +29,7 @@ import java.util.*;
  * @author Viktor Terrier {@literal <viktor.terrier at rte-france.com>}
  */
 @JsonTypeName("aligned-range-action")
-public class AlignedRangeAction extends AbstractRemedialAction<RangeAction> implements RangeAction {
+public class AlignedRangeAction extends AbstractRemedialAction implements RangeAction {
 
     public static final int TEMP_VALUE_ARA = 0;
