+++ data/crac/crac-result-extensions/src/main/java/com/farao_community/farao/data/crac_result_extensions/NetworkActionResult.java
@@ -9,15 +9,33 @@ package com.farao_community.farao.data.crac_result_extensions;
 
 import com.farao_community.farao.data.crac_api.NetworkAction;
 import com.farao_community.farao.data.crac_api.State;
+import com.powsybl.commons.extensions.AbstractExtension;
 
+import java.util.HashMap;
+import java.util.Map;
 import java.util.Set;
 
 /**
  * @author Joris Mancini {@literal <joris.mancini at rte-france.com>}
  */
-public class NetworkActionResult extends AbstractRemedialActionResult<NetworkAction> {
+public class NetworkActionResult extends AbstractExtension<NetworkAction> {
+    protected Map<State, Boolean> activationMap;
+
     public NetworkActionResult(Set<State> states) {
-        super(states);
+        activationMap = new HashMap<>();
+        states.forEach(state -> activationMap.put(state, false));
+    }
+
+    public boolean isActivated(State state) {
+        return activationMap.getOrDefault(state, false);
+    }
+
+    public void activate(State state) {
+        activationMap.put(state, true);
+    }
+
+    public void deactivate(State state) {
+        activationMap.put(state, false);
     }
 
     @Override
