+++ data/crac/crac-result-extensions/src/main/java/com/farao_community/farao/data/crac_result_extensions/RangeActionResult.java
@@ -9,6 +9,7 @@ package com.farao_community.farao.data.crac_result_extensions;
 
 import com.farao_community.farao.data.crac_api.RangeAction;
 import com.farao_community.farao.data.crac_api.State;
+import com.powsybl.commons.extensions.AbstractExtension;
 
 import java.util.HashMap;
 import java.util.Map;
@@ -17,14 +18,12 @@ import java.util.Set;
 /**
  * @author Joris Mancini {@literal <joris.mancini at rte-france.com>}
  */
-public class RangeActionResult extends AbstractRemedialActionResult<RangeAction> {
-    private Map<State, Double> setPointMap;
+public class RangeActionResult extends AbstractExtension<RangeAction> {
+    protected Map<State, Double> setPointMap;
 
     public RangeActionResult(Set<State> states) {
-        super();
         setPointMap = new HashMap<>();
         states.forEach(state -> {
-            activationMap.put(state, false);
             setPointMap.put(state, Double.NaN);
         });
     }
@@ -35,6 +34,10 @@ public class RangeActionResult extends AbstractRemedialActionResult<RangeAction>
 
     public void setSetPoint(State state, double setPoint) {
         setPointMap.put(state, setPoint);
+    }
+
+    public boolean isActivated(State state) {
+        return !setPointMap.getOrDefault(state, Double.NaN).isNaN();
     }
 
     @Override
